Flyttar VirkningstidspunktRequest ut i modellen, sia vi treng denne når vi kallar
